requirements.txt
requests
inquirer
matplotlib
mpld3
pylint
python-dateutil